requirements.txt
google-generativeai
pymupdf
python-docx
streamlit
python-dotenv
Pillow